lead stage

+++ app/src/main/java/net/maxproit/salesforce/masum/activity/lead/LeadStageActivity.java
@@ -282,13 +282,13 @@ public class LeadStageActivity extends BaseActivity {
         else
             myLeadApi.setLoanAmount(0);
         if (interest != null)
-            myLeadApi.setOfferedInterestRate(Integer.valueOf(interest));
+            myLeadApi.setOfferedInterestRate((float)Float.valueOf(interest));
         else
-            myLeadApi.setOfferedInterestRate(0);
+            myLeadApi.setOfferedInterestRate((float) 0);
         if (fee != null)
-            myLeadApi.setOfferedProcessFee(Integer.valueOf(fee));
+            myLeadApi.setOfferedProcessFee((float)Float.valueOf(fee));
         else
-            myLeadApi.setOfferedProcessFee(0);
+            myLeadApi.setOfferedProcessFee((float) 0);
         myLeadApi.setDisbursementDate(DateUtils.getDateFormateForSqlite(disDate));
 
         myLeadApi.setFollowUp(followUp);
